edit result

+++ src/controllers/image.js
@@ -1,4 +1,6 @@
 const imageService = require('../services/image')
+const CustomError = require('../errors/CustomError');
+const codes = require('../errors/code');
 
 const createImage = async (req, res) => {
     const { desc, name, idLesson, url } = req.body
@@ -12,6 +14,9 @@ const editImage = async (req, res) => {
 }
 const deleteImage = async (req, res) => {
     const { id } = req.body
+    if (!id) {
+        throw new CustomError(codes.BAD_REQUEST, 'Hãy nhập id của lesson')
+    }
     const result = await imageService.deleteLessonById(id)
     return res.send(result)
 }
